🐛 [BUGFIX]: FIx glitch when resizing window on Market + big space on clearable element in Search in Market
