+++ .changeset/thick-vans-deny.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": patch
+---
+
+FIx glitch when resizing window on Market + big space on clearable element in Search in Market
