+++ apps/ledger-live-desktop/src/renderer/screens/market/hooks/useMarketCoin.ts
@@ -13,13 +13,14 @@ import { setMarketOptions } from "~/renderer/actions/market";
 import { removeStarredMarketCoins, addStarredMarketCoins } from "~/renderer/actions/settings";
 import { marketParamsSelector } from "~/renderer/reducers/market";
 import { starredMarketCoinsSelector, localeSelector } from "~/renderer/reducers/settings";
+import { useFetchCurrencyAll } from "@ledgerhq/live-common/exchange/swap/hooks/index";
 
 export const useMarketCoin = () => {
   const marketParams = useSelector(marketParamsSelector);
   const { colors } = useTheme();
   const dispatch = useDispatch();
   const { currencyId } = useParams<{ currencyId: string }>();
-
+  const { data: currenciesAll } = useFetchCurrencyAll();
   const starredMarketCoins: string[] = useSelector(starredMarketCoinsSelector);
   const { liveCoinsList, supportedCounterCurrencies } = useMarketDataProvider();
 
@@ -45,6 +46,7 @@ export const useMarketCoin = () => {
     useMarketActions({
       currency: currency,
       page: Page.MarketCoin,
+      currenciesAll,
     });
 
   const color = internalCurrency
